requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
bcrypt==3.2.0
black==20.8b1
blessed==1.17.12
botocore==1.19.34
brazilnum==0.8.8
cached-property==1.5.2
cement==2.8.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.3.1
dataclasses==0.8
Django==3.1.3
djangorestframework==3.12.2
docker==4.4.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==2.10
importlib-metadata==3.1.1
inflect==5.0.2
isort==5.6.4
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
locket==0.2.0
mccabe==0.6.1
mypy-extensions==0.4.3
num2words==0.5.10
numpy==1.19.3
olefile==0.46
packaging==20.4
pandas==1.1.4
paramiko==2.7.2
parso==0.7.1
partd==1.1.0
pathspec==0.5.9
pickleshare==0.7.5
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
quantulum3==0.7.6
regex==2020.11.13
requests==2.23.0
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
stemming==1.0.1
termcolor==1.1.0
texttable==1.6.3
toml==0.10.2
toolz==0.11.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.4
wcwidth==0.1.9
websocket-client==0.57.0
wrapt==1.12.1
zipp==3.4.0
